workout_tracker_api/requirements.txt
python-decouple==3.8
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
# psycopg2-binary==2.9.7
# mysqlclient==2.1.1
whitenoise==6.6.0
